fixed threejs reticle example
After thorough tests I (Iker) determine that prompt has a major performance hit with the call ranging from 1 to 80 milliseconds no matter what the native code executes. Added several modifications to test the prompt performance and to try to improve as much as possible the passing of the pose and projection matrix from the native side. This modifications are far from being final.

All the changes to adapt to the new findings that prompt was not the right way to make performant calls to the native side (although it is synchronous). Adapted some of the examples to handle errors and the new changes (left projection matrix, etc.).

The WebARKitTests folder was removed from the project.

Change-Id: I9767e1abce8aecdf0b9d10bd820962e246133269