Added raf polyfilling. Passed the viewmatrix from the native side to WebVR to get the right values for the ARFrameData instance. Fixed a minor deprecation warning in the metal rendering code. Got rid of the custom matrix to quaternion code and used iOS math libraries instead. Some comments in some specific/important parts of the code. Adapted some examples to use the new viewmatrix both because is more optimal and because needed to be tested.
Fixed some comments to improve the codebase using underscore at the end of the name of a variable to mark it a private in JS, removing some whitespaces and other minor fixes.

Fixed an orientation issue with the projection matrix and renamed some variables to make them more auto explained.

Change-Id: I22e4b8bf9cca115450add7f62500d162a8167845